# perf(serialize): faster `Date` serialization